pyproject.toml
[tool.poetry]
name = "quantlib-ws"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quantlib-python = "^1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
